Fix split package in analysis-phonetic plugin
The analysis-phonetic plugin reuses server the server package name for
analysis. This commit moves the plugin implementation to use a single
package name, o.e.p.analysis.phonetic
