+++ plugins/analysis-phonetic/build.gradle
@@ -27,8 +27,3 @@ restResources {
 tasks.named("dependencyLicenses").configure {
   mapping from: /lucene-.*/, to: 'lucene'
 }
-
-tasks.named('splitPackagesAudit').configure {
-    // Lucene packages should be owned by Lucene!
-  ignoreClasses 'org.elasticsearch.index.analysis.PhoneticTokenFilterFactory'
-}
