+++ plugins/analysis-phonetic/src/test/resources/org/elasticsearch/plugin/analysis/phonetic/phonetic-1.yml
